Crate dynkin

Structs

Queue ğŸ”’

A queue to manage the elements being processed during the closure extension.

Constants

BITSET_WORDS ğŸ”’

MAX_N ğŸ”’

MAX_SUBSETS ğŸ”’

Functions

bs_clear ğŸ”’

Sets all bits in the bitset to zero.

bs_copy ğŸ”’

Copies the contents of one bitset to another.

bs_get ğŸ”’

Gets the value of a specific bit in the bitset. Returns true if the bit is set, false otherwise.

bs_set ğŸ”’

Sets a specific bit in the bitset to 1.

extend_closure ğŸ”’

Extends the closure by including a new element and updating the closure set. Returns false if the extension is not valid (i.e., it conflicts with excluded elements). Parameters:

inner ğŸ”’

Recursively counts the number of valid subsets of the closure. Given a set system F and a property P, the closure of F with respect to P is the smallest superset of F satisfying P. Parameters:

main ğŸ”’

Type Aliases

Bitset ğŸ”’

A bitwise operable representation of the powerset, where 1 represents a valid Dynkin system.